Fhat {splancs}	R Documentation

F nearest neighbour distribution function

Description

Calculates an estimate of the F nearest neighbour distribution function

Usage

Fhat(pts1,pts2,s)

Arguments

pts1
A points data set

pts2
A points data set

s
A vector of distances at which to evaluate Fhat

Details

The function Fhat(pts1,pts2,s) is defined as the proportion of members of a point set pts2 for which the distance to the nearest member of another points set pts1 is less than or equal to s.

Value

A vector of the same length as s, containing the value of Fhat at the distances in s.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

nndistF,Fzero

Examples

data(uganda)
plot(seq(20, 500, 20), Fhat(as.points(uganda), 
as.points(csr(uganda$poly, length(uganda$x))), seq(20, 500, 20)), 
type="l", xlab="distance", ylab="Estimated F")
plot(Ghat(as.points(uganda), seq(20, 500, 20)), Fhat(as.points(uganda), 
as.points(csr(uganda$poly, length(uganda$x))), seq(20, 500, 20)), 
type="l", xlab="Estimated G", ylab="Estimated F")
lines(c(0,1),c(0,1),lty=2)

[Package splancs version 2.01-33 Index]