Ghat {splancs}	R Documentation

G nearest neighbour distribution function

Description

Calculates an estimate of the G nearest neighbour distribution function.

Usage

Ghat(pts,s)

Arguments

pts
A points data set

s
A vector of distances at which to evaluate the G function

Details

The function Ghat(pts,s) is defined as the proportion of members of a point set for which the distance to the nearest other member of the set is less than or equal to s.

Value

A vector of the same length as s, containing the estimate of G at the distances in s.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

Fhat, nndistG

Examples

data(uganda)
plot(seq(20, 500, 20), Ghat(as.points(uganda), seq(20, 500, 20)), 
type="l", xlab="distance", ylab="Estimated G")

[Package splancs version 2.01-34 Index]