bboxx {splancs}	R Documentation

Generate a non-closed bounding polygon

Description

Generate a non-closed bounding polygon from the bounding box of an object

Usage

bboxx(obj)

Arguments

obj
A matrix with two rows and two columns reporting the bounding box of an object

Details

The object used by bboxx may easily be created by using the sp bbox method on an object of interest, such as a points data set.

Value

A points data set of four points giving the non-closed coordinates of the bounding box

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

sbox, bbox

[Package splancs version 2.01-34 Index]