bodmin {splancs}	R Documentation

Bodmin Moors granite tors

Description

Locations of 35 granite tors on Bodmin Moor, taken from Infomap data set (northings multiplied by -1 to correspond to Figure 3.2, p. 82, Bailey and Gatrell.

Usage

data(bodmin)

Format

A list corresponding to a Venables and Ripley point object with 35 observations

x 	 numeric 	 grid eastings

y 	 numeric 	 grid northings

area 	 list 	 bounding box with xl, xu, yl, yu

poly 	 array 	 polygon boundary with columns x and y

Source

Pinder and Witherick, 1977 - Bailey and Gatrell 1995, ch. 3.

References

Bailey, T. C. and Gatrell, A. C. 1995, Interactive spatial data analysis. Longman, Harlow.

[Package splancs version 2.01-39 Index]