delpoints {splancs}	R Documentation

Select points to delete from a points data set

Description

Select points to delete from a points data set.

Usage

delpoints(pts,add=FALSE)

Arguments

pts
a points data set

add
if false, plot the points using pointmap.

Details

Using the mouse, the user selects points on the current graphics device. These points are marked on the plot as they are selected. The function returns the remaining points as a points object. If add is false the points are plotted on the current plot device.

Value

A points object containing the undeleted points.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

[Package splancs version 2.01-34 Index]