dsquare {splancs}	R Documentation

Distance-squared from a number of points to a number of sources

Description

Computes the distance-squared from a number of points to a number of sources.

Usage

dsquare(pts, srcs, namepref="d")

Arguments

pts
A number of points representing the locations of cases and controls.

srcs
A number of points representing source locations

namepref
A prefix given to the name of the results.

Value

A data frame with the same number of columns as srcs. The column names will be the value of namepref prefixing the numbers from 1 to the number of sources.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

tribble, triblik

[Package splancs version 2.01-34 Index]