gen {splancs}	R Documentation

generate points in polygon

Description

generates random points within a defined polygon, trying to reach npoints points - used in csr.

Usage

gen(poly, npoints)

Arguments

poly
A polygon data set

npoints
The number of points to generate

Value

returns a point object.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

csr

[Package splancs version 2.01-34 Index]