gridpts {splancs}	R Documentation

Generate a grid of points

Description

Generate a grid of points

Usage

gridpts(poly,npts,xs,ys)

Arguments

poly
polygon in which to generate the points

npts
approximate number of points to generate

xs,ys
grid spacing in x and y

Either npts or xs and ys must be specified. If all three are given then xs and ys are ignored.

Value

A points object containing a grid of points inside the polygon. If npts is specified, then a grid spacing xs and ys will be calculated to give approximately npts in the polygon. If xs and ys are given then these will be used to generate a number of points in the polygon.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

[Package splancs version 2.01-33 Index]