kernrat {splancs}	R Documentation

Ratio of two kernel smoothings

Description

Return the ratio of two kernel smoothings

Usage

kernrat(pts1,pts2,poly,h1,h2,nx=20,ny=20,kernel='quartic')

Arguments

pts1,pts2
Point data sets

poly
A polygon data set

h1,h2
The kernel width parameters, h1 for pts1, and h2 for pts2

nx
Number of points along the x-axis of the returned grid.

ny
Number of points along the y-axis of the returned grid.

kernel
Type of kernel function to use. Currently only the quartic kernel is implemented.

Value

A list with the following components:

x
List of x-coordinates at which the kernel function has been evaluated.

y
List of y-coordinates at which the kernel function has been evaluated.

z
A matrix of dimension nx by ny containing the ratio of the kernel functions.

h
A vector of length 2 containing h1 and h2

kernel
a character string containing the kernel name.

References

Berman M. and Diggle P.J. (1989) Estimating Weighted Integrals of the Second-Order Intensity of Spatial Point Patterns. J. R. Statist Soc B51 81-92; Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

kernel2d, mse2d

Examples

data(okwhite)
data(okblack)
okpoly <- list(x=c(okwhite$x, okblack$x), y=c(okwhite$y, okblack$y))
kr <- kernrat(as.points(okwhite), as.points(okblack), bboxx(bbox(as.points(okpoly))),
 h1=50, h2=50)
image(kr, asp=1)
brks <- quantile(c(kr$z), seq(0,1,1/10), na.rm=TRUE)
lbrks <- formatC(brks, 3, 6, "g", " ") 
cols <- heat.colors(length(brks)-1)
def.par <- par(no.readonly = TRUE)
layout(matrix(c(1,0,1,2), 2, 2, byrow = TRUE), c(2.5,1.5), c(1,3), TRUE)
image(kr, breaks=brks, col=cols, asp=1)
plot.new()
legend(c(0,1), c(0,1), legend=paste(lbrks[-length(lbrks)], lbrks[-1], sep=":"), fill=cols, bty="n")
par(def.par)

[Package splancs version 2.01-34 Index]