khvc {splancs}	R Documentation

Covariance matrix for the difference between two K-functions

Description

Calculate the covariance matrix for the difference between two K-functions. Also return the contribution to the variance for each of the two point patterns,

Usage

khvc(pts1, pts2, poly, s)

Arguments

pts1
An object containing the case locations.

pts2
An object containing the control locations.

poly
A polygon enclosing the locations in pts1 and pts2

s
A vector of distances at which the calculation is to be made.

Value

A list with four components:

varmat
The upper triangle of the covariance matrix.

k11
The variance of Khat for the cases

k22
The variance of Khat for the controls

k12
The covariance of Khat for the cases and Khat for controls.

Note

Note that the diagonal of the covariance matrix is $k11 - 2 * $k12 + $k22

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

khat, khvmat, secal

[Package splancs version 2.01-34 Index]