khvmat {splancs}	R Documentation

Covariance matrix for the difference between two K-functions

Description

Calculate the covariance matrix for the difference between two K-functions under random labelling of the corresponding two sets of points.

Usage

khvmat(pts1, pts2, poly, s)

Arguments

pts1
An object containing the case locations.

pts2
An object containing the control locations.

poly
Polygon enclosing the points in pts1 and pts2.

s
A vector of distances at which the calculation is to be made.

Value

A matrix containing the covariances, with the variances on the diagonal.

References

Diggle P.J and Chetwynd A.C (1991) Second order analysis of spatial clustering Biometrics 47 1155-63; Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

secal

[Package splancs version 2.01-33 Index]