mpoint {splancs}	R Documentation

Overlay a number of point patterns

Description

Overlay a number of point patterns.

Usage

mpoint(...,cpch,add=FALSE,type="p")

Arguments

...
At least one argument consisting of a points data set must be specified.

cpch
A vector of characters for plotting symbols

add
if add is TRUE then overlay on an existing plot

type
plot data as points if type="p", lines if type="l"

Details

mpoint enables several point or polygon datasets to be overlayed. The plot region is calculated so that all the specified datasets fit in the region. The parameter cpch specifies the characters to use for each set of points. The default cpch consists of the numbers 1 to 9 followed by the uppercase letters A to Z. If cpch is shorter than the number of point sets to plot, then it is repeated.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

pointmap

[Package splancs version 2.01-33 Index]