nndistG {splancs}	R Documentation

Nearest neighbour distances as used by Ghat()

Description

Calculate nearest neighbour distances as used by Ghat().

Usage

nndistG(pts)

Arguments

pts
A points data set

Value

Returns a list with components $dists and $neighs. $dists[i] is the distance to the nearest neighbour of point i in pts, and $neighs[i] is the index of the neighbour of point i.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

nndistF, Fhat, Ghat, Fzero

Examples

data(uganda)
boxplot(nndistG(as.points(uganda))$dists)
plot(ecdf(nndistG(as.points(uganda))$dists))

[Package splancs version 2.01-34 Index]