okblack {splancs}	R Documentation

Oklahoma black offenders

Description

Locations of theft from property offences committed by black offenders in Oklahoma City

Usage

data(okblack)

Format

A list corresponding to a Venables and Ripley point object with 147 observations

x 	 numeric 	 grid eastings

y 	 numeric 	 grid northings

area 	 list 	 bounding box with xl, xu, yl, yu

Source

Carter and Hill, 1979, - Bailey and Gatrell 1995, ch. 3.

References

Bailey, T. C. and Gatrell, A. C. 1995, Interactive spatial data analysis. Longman, Harlow.

[Package splancs version 2.01-39 Index]