pcp {splancs}	R Documentation

Fit a Poisson cluster process

Description

The function fits a Poisson cluster process to point data for a given enclosing polygon and fit parameters

Usage

pcp(point.data, poly.data, h0=NULL, expo=0.25, n.int=20)

Arguments

point.data
a points object

poly.data
a polygon enclosing the study region

h0
upper bound of integration in the criterion function

expo
exponent in the criterion function

n.int
number of intervals used to approximate the integral in the criterion function with a sum

Value

The function returns an object as returned by optim, including:

par
The best set of parameters s2 and rho found

value
The value of the fit corresponding to ‘par’

convergence
‘0’ indicates successful convergence

Author(s)

Giovanni Petris <GPetris@uark.edu>, Roger.Bivand@nhh.no

References

Diggle, P. J. (1983) Statistical analysis of spatial point patterns, London: Academic Press, pp. 55-57 and 78-81; Bailey, T. C. and Gatrell, A. C. (1995) Interactive spatial data analysis, Harlow: Longman, pp. 106-109.

See Also

optim, pcp.sim, Kenv.pcp, khat

Examples

data(cardiff)
polymap(cardiff$poly)
pointmap(as.points(cardiff), add=TRUE)
title("Locations of homes of 168 juvenile offenders")
pcp.fit <- pcp(as.points(cardiff), cardiff$poly, h0=30, n.int=30)
pcp.fit

[Package splancs version 2.01-33 Index]