pip {splancs}	R Documentation

Points inside or outside a polygon

Description

Return points inside or outside a polygon.

Usage

pip(pts,poly,out=FALSE,bound=NULL,quiet=TRUE)

Arguments

pts
A points data set

poly
A polygon data set

out
If out=TRUE, return the points outside the polygon, else the points inside.

bound
If points fall exactly on polygon boundaries, the default NULL gives arbitrary assignments. If TRUE, then all points "on" boundaries are set as within the polygon, if FALSE, outside.

quiet
Do not report which points are on boundary for non-NULL bound

Details

pip calls inout, then selects the appropriate sub-set of points.

Value

pip returns the points of pts that lie inside (or outside with out=TRUE) the polygon poly. Compare this with inpip, which returns the indices of the points in the polygon, and inout which returns a logical vector indicating whether points are inside or outside.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

inpip, inout

[Package splancs version 2.01-33 Index]