print.ribfit {splancs}	R Documentation

Display the fit from tribble()

Description

Display the fit from tribble

Usage

## S3 method for class 'ribfit'
print(x, ...)

Arguments

x
An object returned from tribble

...
optional arguments to pass through to print()

Details

The parameter estimates and log-likelihood for the raised incidence model are displayed. The likelihood ratio, D = 2*(L-Lo), is also given. This function is called whenever print operates on an object with class ribfit.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

tribble

[Package splancs version 2.01-33 Index]