rtor.shift {splancs}	R Documentation

Random toroidal shift on a point data set

Description

Perform a random toroidal shift on a point data set

Usage

rtor.shift(pts,rect)

Arguments

pts
The point data set to shift

rect
A rectangle defining the region for the toroidal map. If not given, the bounding box of pts is used.

Details

The planar region defined by rect is assumed connected at its top and bottom edges, and at its left and right sides. A random shift is applied to the points and the resulting set of points returned.

Value

A point data set like pts, but after application of a random toroidal shift along the x and y axes.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

Shift

[Package splancs version 2.01-34 Index]