sbox {splancs}	R Documentation

Generate a box surrounding a point object

Description

Generate a box surrounding a point object

Usage

sbox(pts, xfrac = .1, yfrac = .1)

Arguments

pts
A points data set

xfrac
The fraction of the width of the point pattern by which the box will surround the point pattern to the left and right.

yfrac
The fraction of the height of the point pattern by which the box will surround the point pattern to the top and bottom.

Value

A points data set of four points giving the coordinates of the surrounding box

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

bboxx

[Package splancs version 2.01-34 Index]