secal {splancs}	R Documentation

Standard errors for the difference between two K-functions

Description

Calculate standard errors for the difference between two K-functions under random labelling of the corresponding two sets of points.

Usage

secal(pts1,pts2,poly,s)

Arguments

pts1,pts2
Two point data sets

poly
Polygon enclosing the points in pts1 and pts2

s
A vector of distances at which to calculate the standard error.

Details

To compare two point patterns, one can calculate the difference between their K-functions. The function secal gives the pointwise standard errors for the estimated differences, under the random labelling hypothesis.

Value

A vector like s containing the value of the standard error at each of the distances in s

References

Diggle P.J. and Chetwynd A.G. (1991) Second-order analysis of spatial clustering Biometrics 47 1155–63; Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

khat

[Package splancs version 2.01-33 Index]