Shift {splancs}	R Documentation

Shift a point data set

Description

Shift a point data set (function name changed from shift to Shift to avoid collision with spatstat)

Usage

Shift(pts,xsh=0.0,ysh=0.0)

Arguments

pts
The point data set to shift

xsh
Amount to shift along the x-axis

ysh
Amount to shift along the y-axis

Value

A point data set like pts, but with xsh added to its x-coordinates, and ysh added to its y-coordinates.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

tor.shift

[Package splancs version 2.01-33 Index]