southlancs {splancs}	R Documentation

Cancer cases in Chorley-Ribble

Description

Locations of cases of cancer of lung and larynx in Chorley-Ribble, Lancashire. The data set is split into a points object southlancs.pts and a case/control 0/1 vector southlancs.cc. There are 917 controls and 57 cases in this data set - these numbers differ from 978 and 58 in Diggle (1990) and Diggle and Rowlingson (1994). The data set also includes the approximate location of an old incinerator old.incinerator, as well as southlancs.bdy, the study area boundary.

Usage

data(southlancs)

Format

A data frame with 974 observations

[,1] 	 x 	 numeric 	 grid eastings (metres)

[,2] 	 y 	 numeric 	 grid northings (metres)

[,3] 	 cc 	 numeric 	 case/control, lung=0, larynx=1

Source

Diggle, Gatrell and Lovett, 1990, - Bailey and Gatrell 1995, ch. 3.

References

Bailey and Gatrell 1995, ch. 3; Diggle, P. (1990) A point process modelling approach to raised incidence of a rare phenomenon in the viscinity of a prespecified point. Journal of the Royal Statistical Society, A, 153, 349-362; Diggle, P. and Rowlingson, B. (1994) A conditional approach to point process modelling of elevated risk. Journal of the Royal Statistical Society, A, 157, 433-440.

Examples

data(southlancs)
op <- par(mfrow=c(2,1))
pointmap(southlancs.pts[southlancs.cc == 0,])
pointmap(old.incinerator, add=TRUE, col="red", pch=19)
title("Lung cancer controls")
pointmap(southlancs.pts[southlancs.cc == 1,])
pointmap(old.incinerator, add=TRUE, col="red", pch=19)
title("Larynx cancer cases")
par(op)
polymap(southlancs.bdy,border="grey")
contour(kernel2d(southlancs.pts[southlancs.cc == 0,], 
	southlancs.bdy, h=500, nx=100, ny=100), nlevels=20, 
	add=TRUE,drawlabels=FALSE)
pointmap(southlancs.pts[southlancs.cc == 1,], add=TRUE, pch=19,
	 col="green")
pointmap(old.incinerator, add=TRUE, pch=19, col="red")
title(xlab="h=500, quartic kernel")
title("Density map of control, green case points, red old incinerator")

[Package splancs version 2.01-34 Index]