splancs {splancs}	R Documentation

Return version number and author information

Description

Return version number and author information

Usage

splancs()

Value

The version string is returned. This is a number of the format x.yy, where x is the major version number and yy is the minor version number.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

[Package splancs version 2.01-33 Index]