stkhat {splancs}	R Documentation

Space-time K-functions

Description

Compute the space-time K-functions

Usage

stkhat(pts, times, poly, tlimits, s, tm)

Arguments

pts
A set of points as defined in Splancs

times
A vector of times, the same length as the number of points in pts

poly
A polygon enclosing the points

tlimits
A vector of length 2 specifying the upper and lower temporal domain.

s
A vector of spatial distances for the analysis.

tm
A vector of times for the analysis

Value

A list with the following components is returned:

s, t
The spatial and temporal scales

ks
The spatial K-function

kt
The temporal K-function

kst
The space-time K-function

For details see Diggle, Chetwynd, Haggkvist, and Morris (1995)

References

Diggle, P., Chetwynd, A., Haggkvist, R. and Morris, S. 1995 Second-order analysis of space-time clustering. Statistical Methods in Medical Research, 4, 124-136;Bailey, T. C. and Gatrell, A. C. 1995, Interactive spatial data analysis. Longman, Harlow, pp. 122-125; Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

stsecal, stvmat, stmctest, stdiagn

Examples

data(burkitt)
bur1 <- stkhat(burpts, burkitt$t, burbdy, c(400, 5800),
  seq(1,40,2), seq(100, 1500, 100))
oldpar <- par(mfrow=c(2,1))
plot(bur1$s, bur1$ks, type="l", xlab="distance", ylab="Estimated K",
  main="spatial K function")
plot(bur1$t, bur1$kt, type="l", xlab="time", ylab="Estimated K",
  main="temporal K function")
par(oldpar)

[Package splancs version 2.01-34 Index]