tribble {splancs}	R Documentation

Diggle-Rowlingson Raised Incidence Model

Description

Fits the Diggle-Rowlingson Raised Incidence Model.

Usage

tribble(ccflag, vars=NULL, alphas=NULL, betas=NULL, rho, 
 which=1:length(alphas), covars=NULL, thetas=NULL, 
 steps=NULL, reqmin=0.001, icount=50, hessian=NULL)

Arguments

ccflag
Case-control flag : a vector of ones and zeroes.

vars
A matrix where vars[i,j] is the distance squared from point i to source j.

alphas
Initial value of the alpha parameters.

betas
Initial value of the beta parameters.

rho
Initial value of the rho parameter.

which
Defines the mapping from sources to parameters.

covars
A matrix of covariates to be modelled as log-linear terms. The element covars[i,j] is the value of covariate j for case/control i.

thetas
Initial values of covariate parameters.

steps
Step sizes for the Nelder-Mead simplex algorithm.

reqmin
Tolerance for simplex algorithm

icount
Iteration count for simplex algorithm

hessian
by default NULL, any other value causes hessian to be computed and returned

Value

The return value is a list with many components, and class ribfit.

alphas
A vector of the alpha parameters at the maximum

betas
A vector of the beta values at the maximum

rho
The value of rho at the maximum

logl
The maximised log-likelihood

null.logl
The null log-likelihood

call
The function call to tribble

For further information see Diggle and Rowlingson (1993).

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

triblik, dsquare

[Package splancs version 2.01-33 Index]