triblik {splancs}	R Documentation

Log-likelihood for the Diggle-Rowlingson raised incidence model

Description

Calculates the log-likelihood for the Diggle-Rowlingson raised incidence model.

Usage

triblik(ccflag, vars=NULL, alphas=NULL, betas=NULL, rho, 
 which=1:length(alphas), covars=NULL, thetas=NULL)

Arguments

ccflag
Case-control flag : a vector of ones and zeroes.

vars
A matrix where vars[i,j] is the distance squared from point i to source j.

alphas
The alpha parameters.

betas
The beta parameters.

rho
The rho parameter.

which
Defines the mapping from sources to parameters.

covars
A matrix of covariates to be modelled as log-linear terms. The element covars[i,j] is the value of covariate j for case/control i.

thetas
The covariate parameters.

Value

The log-likelihood for the given parameters and the given distances and optional covariates is returned.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

tribble, dsquare

[Package splancs version 2.01-34 Index]