uganda {splancs}	R Documentation

Craters in Uganda

Description

Locations of craters in a volcanic field in Uganda

Usage

data(uganda)

Format

A list corresponding to a Venables and Ripley point object with 120 observations

x 	 numeric 	 grid eastings

y 	 numeric 	 grid northings

area 	 list 	 bounding box with xl, xu, yl, yu

poly 	 array 	 polygon boundary with columns x and y

Source

Tinkler, 1971, - Bailey and Gatrell 1995, ch. 3.

References

Bailey, T. C. and Gatrell, A. C. 1995, Interactive spatial data analysis. Longman, Harlow.

[Package splancs version 2.01-33 Index]