zoom {splancs}	R Documentation

Interactively specify a region of a plot for expansion

Description

Interactively specify a region of a plot for expansion

Usage

zoom(quiet=FALSE,out=FALSE,...)

Arguments

quiet
If false, prompt the user to enter two coordinates. If true, say nothing.

out
If true, expand the limits of the current plot by a factor of three, centred on the current plot.

...
Other arguments are passed through to pointmap.

Details

A prompt is optionally displayed, and the user selects two points forming the diagonal of a rectangle. A new, empty plot is created that has its axis limits set to the bounding square of the selected rectangle. If out=TRUE, no prompt is displayed, and a new blank plot is created with its limits in x and y set to span an area three times the height and width centred on the current centre.

Value

None

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

pointmap

[Package splancs version 2.01-34 Index]