as.points {splancs}	R Documentation

Creates data in spatial point format

Description

Creates data in spatial point format.

Usage

as.points(...)

Arguments

...
any object(s), such as x and y vectors of the same length, or a list or data frame containing x and y vectors. Valid options for ... are: a points object ; returns it unaltered; a list with x and y elements of the same length — returns a points object with the x and y elements as the coordinates of the points; two vectors of equal length ; returns a points object with the first vector as the x coordinates, the second vector as the y-coordinates.

Value

as.points tries to return the argument(s) as a points object.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

[Package splancs version 2.01-39 Index]