Select points inside a polygon

Select points inside a polygon

inpip(pts,poly,bound=NULL,quiet=TRUE)

Arguments

pts

A points data set

poly

A polygon data set

bound

If points fall exactly on polygon boundaries, the default NULL gives arbitrary assignments. If TRUE, then all points "on" boundaries are set as within the polygon, if FALSE, outside.

quiet

Do not report which points are on boundary for non-NULL bound

Value

inpip returns a vector of indices of the points in pts that are located in the polygon. Note that "in" is an arbitrary concept for points "on" the polygon boundary.

See also

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: https://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.