nndistF {splancs}	R Documentation

Nearest neighbour distances as used by Fhat()

Description

Calculate nearest neighbour distances as used by Fhat()

Usage

nndistF(pts1,pts2)

Arguments

pts1
A points data set

pts2
A points data set

Value

The set of distances from each of the points in pts2 to the nearest point in pts1 is returned as a vector.

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: http://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.

See Also

nndistG, Fhat, Ghat, Fzero

Examples

data(uganda)
boxplot(nndistF(as.points(uganda), as.points(csr(uganda$poly, length(uganda$x)))))
plot(ecdf(nndistF(as.points(uganda), 
as.points(csr(uganda$poly, length(uganda$x))))),
main="Fhat ecdf Uganda volcano data")

[Package splancs version 2.01-39 Index]