Point Objects

Creates and tests for data in spatial point format.

spoints(data,npoints)

Arguments

data

vector containing the data values for the points in order (x1,y1),(x2,y2),...

npoints

number of points to generate, if missing, set to length(data)/2.

Value

spoints returns an object suitable for use as a point data object. If npoints is given, the vector data is either truncated or repeated until sufficient data values are generated. The returned object is a two-column matrix, where the first column stores the x-coordinate, and the second column stores the y-coordinate.

See also

References

Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: https://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.