Variance matrix for space-time clustering

Compute the variance matrix for space-time clustering

stvmat(pts, times, poly, tlim, s, tm)

Arguments

pts

A set of points.

times

A vector of times, the same length as the number of points in pts

poly

A polygon that encloses the points

tlim

A vector of length 2 specifying the upper and lower temporal domain.

s

A vector of spatial distances for the analysis

tm

A vector of times for the analysis

Value

A four-dimensional matrix is returned. The covariance between space-time t1,s1 and t2,s2 is given by the corresponding element [t1,s1,t2,s2]

For full details, see Diggle, Chetwynd, Haggkvist and Morris (1995)

See also

References

Diggle, P., Chetwynd, A., Haggkvist, R. and Morris, S. 1995 Second-order analysis of space-time clustering. Statistical Methods in Medical Research, 4, 124-136; Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655; the original sources can be accessed at: https://www.maths.lancs.ac.uk/~rowlings/Splancs/. See also Bivand, R. and Gebhardt, A. 2000 Implementing functions for spatial statistical analysis using the R language. Journal of Geographical Systems, 2, 307-317.